refactor: TransactionTemplate 적용
