+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -2,9 +2,8 @@ package com.techcourse.service;
 
 import com.techcourse.config.DataSourceConfig;
 import com.techcourse.domain.User;
-import java.sql.SQLException;
-import org.springframework.dao.DataAccessException;
-import org.springframework.jdbc.datasource.DataSourceUtils;
+
+import static org.springframework.transaction.support.TransactionTemplate.execute;
 
 public class TxUserService implements UserService {
 
@@ -16,32 +15,24 @@ public class TxUserService implements UserService {
 
     @Override
     public User findById(final long id) {
-        return userService.findById(id);
+        return execute(DataSourceConfig.getInstance(), () -> userService.findById(id));
     }
 
     @Override
     public void insert(final User user) {
-        userService.insert(user);
+        execute(DataSourceConfig.getInstance(), () -> {
+            userService.insert(user);
+            return null;
+        });
     }
 
     @Override
     public void changePassword(final long id, final String newPassword, final String createBy) {
         final var dataSource = DataSourceConfig.getInstance();
-        final var connection = DataSourceUtils.getConnection(dataSource);
-        try {
-            connection.setAutoCommit(false);
+        execute(dataSource, () -> {
             userService.changePassword(id, newPassword, createBy);
-
-            connection.commit();
-        } catch (final Exception e) {
-            try {
-                connection.rollback();
-            } catch (SQLException ex) {
-                throw new DataAccessException(ex);
-            }
-            throw new DataAccessException(e);
-        } finally {
-            DataSourceUtils.releaseConnection(connection, dataSource);
-        }
+            return null;
+        });
     }
+
 }
