+++ jdbc/src/main/java/org/springframework/jdbc/core/JdbcTemplate.java
@@ -34,30 +34,8 @@ public class JdbcTemplate {
         return results.get(0);
     }
 
-    public <T> T queryForObject(final Connection connection, final String sql, final RowMapper<T> rowMapper, final Object... params) {
-        List<T> results = query(connection, sql, rowMapper, params);
-        if (results.size() > 1) {
-            throw new DataAccessException("too many result. expected 1 but was " + results.size());
-        }
-        if (results.isEmpty()) {
-            throw new DataAccessException("no result");
-        }
-        return results.get(0);
-    }
-
     public <T> List<T> query(final String sql, final RowMapper<T> rowMapper, final Object... parameters) {
         final var conn = getConnection();
-        try (final PreparedStatement preparedStatement = conn.prepareStatement(sql);
-             final ResultSet resultSet = executeQuery(preparedStatement, parameters)) {
-            log.debug("query : {}", sql);
-            return mapResults(rowMapper, resultSet);
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new DataAccessException(e);
-        }
-    }
-
-    public <T> List<T> query(final Connection conn, final String sql, final RowMapper<T> rowMapper, final Object... parameters) {
         try (final PreparedStatement preparedStatement = conn.prepareStatement(sql);
              final ResultSet resultSet = executeQuery(preparedStatement, parameters)) {
             log.debug("query : {}", sql);
@@ -91,18 +69,6 @@ public class JdbcTemplate {
 
     public void update(final String sql, final Object... parameters) {
         final Connection conn = getConnection();
-        try (final PreparedStatement preparedStatement = conn.prepareStatement(sql)) {
-            log.debug("query : {}", sql);
-
-            setParameters(preparedStatement, parameters);
-            preparedStatement.executeUpdate();
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new DataAccessException(e);
-        }
-    }
-
-    public void update(final Connection conn, final String sql, final Object... parameters) {
         try (final PreparedStatement preparedStatement = conn.prepareStatement(sql)) {
             log.debug("query : {}", sql);
