+++ jdbc/src/main/java/org/springframework/transaction/support/TransactionExecution.java
@@ -1,0 +1,5 @@
+package org.springframework.transaction.support;
+
+public interface TransactionExecution<T> {
+    T execute();
+}
